refactor: merge CheckTypoToleranceDomainService into concrete @DomainService class
